backend/InvoicingPython/requirements.txt
alembic==1.10.3
amqp==5.1.1
aniso8601==9.0.1
attrs==23.1.0
billiard==3.6.4.0
cattrs==22.2.0
celery==5.2.7
Cerberus==1.3.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
exceptiongroup==1.1.1
Flask==2.2.3
Flask-CeleryExt==0.5.0
flask-expects-json==1.7.0
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
pika==1.3.1
prompt-toolkit==3.0.38
psycopg2==2.9.6
pycparser==2.21
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
